# api_documentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Basic Bank System - OpenAPI 3.0",
    "description": "Dokumentasi API Basic Bank System\n\n- [Repository of Basic Bank System](https://github.com/yusuftri023/Basic-Banking-System)",
    "contact": {
      "email": "yusuftri023@gmail.com"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
...
  "paths": {
    "/register": {
      "post": {
        "tags": ["Auth (Token Based Authentication)"],
        "summary": "Create a new account for the website",
        "description": "Create a new account for the website",
        "operationId": "registerForm",
        "parameters": [
          {
            "name": "json-web-token",
            "in": "header",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "request form for registration",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "balance": {
                    "type": "string",
                    "example": 20000000
                  },
                  "password": {
                    "type": "string",
                    "example": "eu28219f212ej43912k!#&!@0"
                  },
                  "email": {
                    "type": "string",
                    "example": "hendraww23@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Account has been created"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "account_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "balance": {
                          "type": "integer",
                          "example": 100000000
                        },
                        "email": {
                          "type": "string",
                          "example": "hendrawww@gmail.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Auth (Token Based Authentication)"],
        "summary": "Login to authenticate and create json web token",
        "description": "Login to authenticate and create json web token",
        "operationId": "loginAccount",
        "requestBody": {
          "description": "Login and authenticate user by giving json web token",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "example": 12345678
                  },
                  "email": {
                    "type": "string",
                    "example": "yusufhaha@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Login success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accountData": {
                          "type": "object",
                          "properties": {
                            "account_id": {
                              "type": "integer",
                              "example": 25
                            },
                            "account_number": {
                              "type": "integer",
                              "example": 25
                            },
                            "balance": {
                              "type": "integer",
                              "example": 1000000
                            },
                            "user_id": {
                              "type": "integer",
                              "example": 5
                            },
                            "email": {
                              "type": "string",
                              "example": "yusufhaha@gmail.com"
                            },
                            "createAt": {
                              "type": "string",
                              "example": "2023-11-07T14:08:35.000Z"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2023-11-07T07:08:35.253Z"
                            }
                          }
                        },
                        "token": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50X2lkIjozNywiYWNjb3VudF9udW1iZXIiOjM3LCJiYWxhbmNlIjoiMTAwMDAwIiwidXNlcl9pZCI6NSwiZW1haWwiOiJqYW1AZ21haWwuY29tIiwiY3JlYXRlQXQiOiIyMDIzLTExLTA3VDE0OjA4OjM1LjAwMFoiLCJ1cGRhdGVkQXQiOiIyMDIzLTExLTA3VDA3OjA4OjM1LjI1M1oiLCJpYXQiOjE2OTkzNTA5MTEsImV4cCI6MTY5OTM1NDUxMX0.lbzDZdEdNOceKx-e82zT5A8lW15b4OrN_iD3-UwE-RU"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Login failed : Your email or password does not match"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get Accounts lists",
        "description": "Get Accounts Lists",
        "operationId": "getAccounts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination page",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination display limit",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 8887329342
                    },
                    "message": {
                      "type": "string",
                      "example": "User data succesfully added to the database"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "account_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "balance": {
                          "type": "integer",
                          "example": 1000000
                        },
                        "email": {
                          "type": "string",
                          "example": "hendraww23@gmail.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed operation should have been 200 but open api won't let response code duplicate exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Data Empty"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "Add a new account for user to database",
        "description": "Add a new account for user to database",
        "operationId": "Accounts",
        "requestBody": {
          "description": "Create a new Users in the website",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "balance": {
                    "type": "string",
                    "example": 20000000
                  },
                  "password": {
                    "type": "string",
                    "example": "eu28219f212ej43912k!#&!@0"
                  },
                  "email": {
                    "type": "string",
                    "description": "Serial number of identity card",
                    "example": "hendraww23@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Account has been created"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "account_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "balance": {
                          "type": "integer",
                          "example": 100000000
                        },
                        "email": {
                          "type": "string",
                          "example": "hendrawww@gmail.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "name": "access_token",
        "in": "cookie"
      }
    }
  }
}